Move ShoppingSite to a standalone file
